refactor(#462): split out process_spec and feature_spec
Fixes #462
