+++ requirements.lock
@@ -52,6 +52,8 @@ pytz==2024.1
     # via pandas
 rich==13.7.0
     # via timeseriesflattener
+ruff==0.2.2
+    # via timeseriesflattener
 scikit-learn==1.4.1.post1
     # via timeseriesflattener
 scipy==1.12.0
