+++ src/test_benchmark.py
@@ -8,14 +8,10 @@ import numpy as np
 import polars as pl
 import pytest
 from iterpy.iter import Iter
-from timeseriesflattenerv2.aggregators import MaxAggregator, MeanAggregator
-from timeseriesflattenerv2.feature_specs import (
-    Aggregator,
-    LookDistance,
-    PredictionTimeFrame,
-    PredictorSpec,
-    ValueFrame,
-)
+from timeseriesflattenerv2.aggregators import Aggregator, MaxAggregator, MeanAggregator
+from timeseriesflattenerv2.feature_specs.meta import LookDistance, ValueFrame
+from timeseriesflattenerv2.feature_specs.prediction_times import PredictionTimeFrame
+from timeseriesflattenerv2.feature_specs.predictor import PredictorSpec
 from timeseriesflattenerv2.flattener import Flattener
