+++ src/timeseriesflattenerv2/feature_specs/default_column_names.py
@@ -1,0 +1,4 @@
+default_entity_id_col_name = "entity_id"
+default_pred_time_uuid_col_name = "pred_time_uuid"
+default_pred_time_col_name = "pred_timestamp"
+default_timestamp_col_name = "timestamp"
